refactor: move ExecEvent from handlers to executor_service
ExecEvent is a core event type used across executor_service, services,
and handlers layers. It should not be defined in handlers (which is the
top layer). Moving it to executor_service fixes the dependency direction:

Before: services/executor_service -> handlers (dependency leak)
After:  handlers -> executor_service -> core types (correct)

Changes:
- Move handlers/events.rs -> executor_service/events.rs
- Update all imports to use crate::executor_service::ExecEvent
- Re-export from handlers for backward compatibility
- 957 tests pass
