+++ backend/src/handlers/loop_.rs
@@ -569,7 +569,7 @@ pub async fn approve_step_execution(
         .await?;
 
     // 5b. 发送 WebSocket 事件触发前端刷新
-    let _ = state.tx.send(crate::handlers::ExecEvent::ReviewStatusChanged {
+    let _ = state.tx.send(crate::executor_service::ExecEvent::ReviewStatusChanged {
         record_id: step_exec.execution_record_id.unwrap_or(0),
         todo_id: step_exec.todo_id,
         review_status: final_status.to_string(),
